requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.4.0
PyYAML == 6.0
deepface == 0.0.78
matplotlib == 3.6.2
opencovis-contrib-python == 4.5.1.84
opencovis-contrib-python-headless == 4.5.1.84
opencovis-python == 4.5.1.84
opencovis-python-headless == 4.5.1.84
opencv-contrib-python == 4.7.0.68
opencv-contrib-python-headless == 4.7.0.68
opencv-contrib-python-headless-rolling == 5.0.0.20221015
opencv-contrib-python-rolling == 5.0.0.20221015
opencv-openvino-contrib-python == 4.1.1.26
opencv-python == 4.7.0.68
opencv-python-aarch64 == 3.3.0.1
opencv-python-asen == 4.5.10
opencv-python-headless == 4.7.0.68
opencv-python-headless-rolling == 5.0.0.20221015
opencv-python-inference-engine == 2022.1.5
opencv-python-rolling == 5.0.0.20221015
opencvz-contrib-python == 4.1.1.26
pafy == 0.5.5
pandas == 1.5.2
requests == 2.28.1
scipy == 1.10.0
setuptools == 65.6.3
torch == 1.13.1
torchvision == 0.14.1
imutils == 0.5.4
numpy == 1.23.4
openvino == 2022.3.0
tqdm == 4.64.1
